move e2e tests into their own files
