+++ package.json
@@ -13,7 +13,8 @@
     "format": "prettier --write 'src/**/*.ts'",
     "format:check": "prettier --check 'src/**/*.ts'",
     "lint": "eslint 'src/**/*.ts'",
-    "test": "NODE_OPTIONS=--experimental-vm-modules jest"
+    "test": "NODE_OPTIONS=--experimental-vm-modules jest",
+    "baselines:update": "NODE_OPTIONS=--experimental-vm-modules UPDATE_MODE=true jest src/test/*.e2e.ts"
   },
   "bin": {
     "literate-ts": "./bin/literate-ts.js"
